requirements.txt
sentence-transformers==2.2.2
pinecone-client==3.2.2
python-docx==1.1.0
google-generativeai==0.4.1
numpy>=1.21.0
python-dotenv>=1.0.0
